ver 0.2.07 changing format of mqdump for mqbyte24
git-svn-id: http://syspver1/svn/develop/modules/mqUtil@1686 72f11a4d-5a25-4e8a-b8c8-85b0af744ad3

+++ src/mqdump.c
@@ -786,7 +786,7 @@ void dumpMqBag( MQHBAG bag )
       goto _door;                                    //
     }                                                //
   }                                                  //
-  logger( LMQM_ITEM_COUNT, itemCount );              //
+  logger( LMQM_ITEM_COUNT, (int) itemCount );        //
                                                      //
   // -------------------------------------------------------
   // go through all items
